requirements.txt
discord.py
python-dotenv
feedparser
markdownify
psutil
ffmpeg4discord
aiohttp
asyncio